split Reducer.js to smaller individual reducers
